+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 12.67.0
+
+* `[usage]` Added support of raw version printing
+* `[path]` Added method `JoinSecure` - more "secure" alternative to standard `Join`
+* `[usage]` Code refactoring
+
 ### 12.66.0
 
 * `[options]` Added method `Arguments.Flatten` for converting arguments into a string